Renamed validator.js to common-validator.js.
